+++ bugowl/api/urls.py
@@ -24,7 +24,8 @@ urlpatterns = [
 	path('agent/sys/bop/', admin.site.urls),
 	path('agent/health_api/', HealthCheckView.as_view(), name='health_api'),
 	path('agent/health_celery/', CeleryHealthCheckView.as_view(), name='health_celery'),
-	path('agent/testcase/', include('testcase.urls')),
+	path('agent/job/', include('job.urls')),
+	# path('agent/testcase/', include('testcase.urls')),
 	# path("agent/testask/", include("testask.urls")),
 	# path("agent/teststep/", include("teststep.urls")),
 ]
